Remove join hack and rename to joinall.py (#196)
* Update join.py

* fixes

* Update and rename join.py to joinall.py
